+++ server.js
@@ -1,24 +1,17 @@
 const express = require('express')
 const connectDB = require('./src/config/db')
 const bodyParser = require('body-parser')
-const authRoutes = require('./src/routes/signup')
-const loginRoutes = require('./src/routes/login')
-const mainRoutes = require('./src/routes/main')
+const routes = require('./src/routes')
+
 require('dotenv').config()
 
 const app = express()
 
 connectDB()
 
-// 미들웨어 설정
 app.use(bodyParser.json())
+app.use('/api', routes)
 
-// 라우터 설정
-app.use('/api/auth', authRoutes)
-app.use('/api/auth', loginRoutes)
-app.use('/api/main', mainRoutes)
-
-// 서버 시작
 const port = process.env.PORT || 3000
 app.listen(port, () => {
   console.log(`Server is running on port ${port}`)
