+++ src/routes/main.js
@@ -1,69 +1,10 @@
 const express = require('express')
-const Tree = require('../models/Tree.js')
-
 const router = express.Router()
 
-// GET /api/main/answer-count
-router.get('/answer-count', async (req, res) => {
-  try {
-    const { user_id } = req.query
+const treeInfoRoutes = require('./main/treeInfo')
+const answerCountRoutes = require('./main/answerCount')
 
-    if (!user_id) {
-      return res.status(400).json({ error: 'user_id가 필요합니다.' })
-    }
-
-    console.log('요청된 user_id:', user_id)
-
-    const tree = await Tree.findOne({ user_id })
-
-    if (!tree) {
-      return res.status(404).json({ error: '성장중인 나무를 찾을 수 없습니다.' })
-    }
-
-    res.json({ answer_count: tree.answer_count })
-  } catch (err) {
-    console.error(err)
-    res.status(500).json({ error: '서버 오류' })
-  }
-})
-// GET /api/main/tree-info
-router.get('/tree-info', async (req, res) => {
-  try {
-    const { user_id } = req.query
-
-    if (!user_id) {
-      return res.status(400).json({ error: 'user_id가 필요합니다.' })
-    }
-
-    console.log('요청된 user_id:', user_id)
-
-    const tree = await Tree.findOne({ user_id })
-
-    if (!tree) {
-      return res.status(404).json({ error: '성장중인 나무를 찾을 수 없습니다.' })
-    }
-
-    let tree_status
-    if (tree.answer_count >= 0 && tree.answer_count <= 4) {
-      tree_status = 1
-    } else if (tree.answer_count >= 5 && tree.answer_count <= 8) {
-      tree_status = 2
-    } else if (tree.answer_count >= 9 && tree.answer_count <= 12) {
-      tree_status = 3
-    } else if (tree.answer_count >= 13 && tree.answer_count <= 16) {
-      tree_status = 4
-    } else {
-      tree_status = '단계 없음'
-    }
-
-    res.json({
-      tree_type: tree.tree_type,
-      tree_status: tree_status,
-    })
-  } catch (err) {
-    console.error(err)
-    res.status(500).json({ error: '서버 오류' })
-  }
-})
+router.use('/tree-info', treeInfoRoutes) 
+router.use('/answer-count', answerCountRoutes) 
 
 module.exports = router
